moved google maps to its own component for simplicity
